// src/generated/models/Cia_agentsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cia_agentsescia_toolused = {
  767150000: 'AgentBuilder',
  767150001: 'CopilotStudio'
} as const;
export type Cia_agentsescia_toolused = keyof typeof Cia_agentsescia_toolused;
export const Cia_agentsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cia_agentsesstatecode = keyof typeof Cia_agentsesstatecode;
export const Cia_agentsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cia_agentsesstatuscode = keyof typeof Cia_agentsesstatuscode;

export interface Cia_agentsesBase {
  cia_agentname: string;
  cia_agentsid: string;
  cia_connectors?: string;
  cia_description?: string;
  cia_linktoagent?: string;
  cia_toolused?: Cia_agentsescia_toolused;
  cia_upvotes?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cia_agentsesstatecode;
  statuscode?: Cia_agentsesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Cia_agentses extends Cia_agentsesBase {
  cia_toolusedname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}