// src/generated/models/Cia_commentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cia_commentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cia_commentsstatecode = keyof typeof Cia_commentsstatecode;
export const Cia_commentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cia_commentsstatuscode = keyof typeof Cia_commentsstatuscode;

export interface Cia_commentsBase {
  "cia_Agent@odata.bind"?: string;
  cia_comment?: string;
  cia_commentid: string;
  cia_commentname: string;
  "cia_User@odata.bind"?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cia_commentsstatecode;
  statuscode?: Cia_commentsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Cia_comments extends Cia_commentsBase {
  cia_agentname?: string;
  cia_username?: string;
  cia_useryominame?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  cia_agent?: object;
  _cia_agent_value?: string;
  cia_user?: object;
  _cia_user_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}