// src/generated/models/Cia_upvotesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cia_upvotesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cia_upvotesesstatecode = keyof typeof Cia_upvotesesstatecode;
export const Cia_upvotesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cia_upvotesesstatuscode = keyof typeof Cia_upvotesesstatuscode;

export interface Cia_upvotesesBase {
  "cia_Agent@odata.bind"?: string;
  cia_name: string;
  cia_upvotesid: string;
  "cia_User@odata.bind"?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cia_upvotesesstatecode;
  statuscode?: Cia_upvotesesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Cia_upvoteses extends Cia_upvotesesBase {
  cia_agentname?: string;
  cia_username?: string;
  cia_useryominame?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  cia_agent?: object;
  _cia_agent_value?: string;
  cia_user?: object;
  _cia_user_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}